refactor: update package structure for TaskStateMapper
