+++ src/main/kotlin/data/datasource/mapper/taskState/TaskStateMapper.kt
@@ -1,6 +1,5 @@
-package org.baghdad.data.datasource.mapper.state
+package data.datasource.mapper.taskState
 
-import data.datasource.mapper.taskState.TaskStateColumns
 import org.baghdad.data.datasource.CsvMapper
 import org.baghdad.logic.model.entities.TaskStateEntity
 import java.util.UUID
