+++ src/test/kotlin/data/datasource/mapper/TaskStateMapperTest.kt
@@ -1,7 +1,7 @@
 package data.datasource.mapper
 
 import com.google.common.truth.Truth.assertThat
-import org.baghdad.data.datasource.mapper.state.TaskStateMapper
+import data.datasource.mapper.taskState.TaskStateMapper
 import org.baghdad.logic.model.entities.TaskStateEntity
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.assertThrows
